requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
flask-cors==3.0.10 